pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-swarm"
version = "0.1.0"
description = "Unified AI Swarm Platform for Certification Engine, Living Dossiers, and AI Lab Observatory"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "core", "data", "graphs", "connectors", "eval", "publish", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
